Revert "Organise admin commands package"
This reverts commit 97bba105e78fe764230fe80360cd1e4582f2a086.
