--- data_safe_haven/commands/admin/register_users.py
@@ -1,74 +1,0 @@
-from typing import Annotated
-
-import typer
-
-from data_safe_haven.administration.users import UserHandler
-from data_safe_haven.config import Config, DSHPulumiConfig
-from data_safe_haven.context import ContextSettings
-from data_safe_haven.exceptions import DataSafeHavenError
-from data_safe_haven.external import GraphApi
-from data_safe_haven.functions import sanitise_sre_name
-from data_safe_haven.utility import LoggingSingleton
-
-from . import admin_command_group
-
-
-@admin_command_group.command(
-    help="Register existing users with a deployed Secure Research Environment."
-)
-def register_users(
-    usernames: Annotated[
-        list[str],
-        typer.Option(
-            "--username",
-            "-u",
-            help="Username of a user to register with this SRE. [*may be specified several times*]",
-        ),
-    ],
-    sre: Annotated[
-        str,
-        typer.Argument(
-            help="The name of the SRE to add the users to.",
-        ),
-    ],
-) -> None:
-    """Register existing users with a deployed SRE"""
-    context = ContextSettings.from_file().assert_context()
-    config = Config.from_remote(context)
-    pulumi_config = DSHPulumiConfig.from_remote(context)
-
-    shm_name = context.shm_name
-    # Use a JSON-safe SRE name
-    sre_name = sanitise_sre_name(sre)
-
-    try:
-        # Check that SRE option has been provided
-        if not sre_name:
-            msg = "SRE name must be specified."
-            raise DataSafeHavenError(msg)
-        LoggingSingleton().info(
-            f"Preparing to register {len(usernames)} user(s) with SRE '{sre_name}'"
-        )
-
-        # Load GraphAPI
-        graph_api = GraphApi(
-            tenant_id=config.shm.entra_tenant_id,
-            default_scopes=["Group.ReadWrite.All", "GroupMember.ReadWrite.All"],
-        )
-
-        # List users
-        users = UserHandler(context, config, pulumi_config, graph_api)
-        available_usernames = users.get_usernames_entra_id()
-        usernames_to_register = []
-        for username in usernames:
-            if username in available_usernames:
-                usernames_to_register.append(username)
-            else:
-                LoggingSingleton().error(
-                    f"Username '{username}' does not belong to this Data Safe Haven deployment."
-                    " Please use 'dsh users add' to create it."
-                )
-        users.register(sre_name, usernames_to_register)
-    except DataSafeHavenError as exc:
-        msg = f"Could not register users from Data Safe Haven '{shm_name}' with SRE '{sre_name}'.\n{exc}"
-        raise DataSafeHavenError(msg) from exc
